GMS/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
pillow
dj-rest-auth
dj-rest-auth[with_social]